Make UsnJrnl APIs internal
